+++ test/configure.test.ts
@@ -8,8 +8,8 @@ import * as assertEx from './assertEx';
 import * as vscode from 'vscode';
 import * as fse from 'fs-extra';
 import * as path from 'path';
-import { Platform, OS } from "../configureWorkspace/config-utils";
 import { ext } from '../extensionVariables';
+import { PlatformOS, Platform } from '../utils/platform';
 import { Suite } from 'mocha';
 import { configure, ConfigureTelemetryProperties, ConfigureApiOptions } from '../configureWorkspace/configure';
 import { TestUserInput, IActionContext, TelemetryProperties } from 'vscode-azureextensionui';
@@ -380,7 +380,7 @@ suite("Configure (Add Docker files to Workspace)", function (this: Suite): void 
     const outputChannel: vscode.OutputChannel = vscode.window.createOutputChannel('Docker extension tests');
     ext.outputChannel = outputChannel;
 
-    async function testDotNetCoreConsole(os: OS, projectFolder: string, projectFileName: string, projectFileContents: string, expectedDockerFileContents?: string): Promise<void> {
+    async function testDotNetCoreConsole(os: PlatformOS, projectFolder: string, projectFileName: string, projectFileContents: string, expectedDockerFileContents?: string): Promise<void> {
         await writeFile(projectFolder, projectFileName, projectFileContents);
         await writeFile(projectFolder, 'Program.cs', dotnetCoreConsole_ProgramCsContents);
 
@@ -402,7 +402,7 @@ suite("Configure (Add Docker files to Workspace)", function (this: Suite): void 
         }
     }
 
-    async function testAspNetCore(os: OS, hostOs: OS, hostOsRelease: string, projectFolder: string, projectFileName: string, projectFileContents: string, expectedDockerFileContents?: string): Promise<void> {
+    async function testAspNetCore(os: PlatformOS, hostOs: PlatformOS, hostOsRelease: string, projectFolder: string, projectFileName: string, projectFileContents: string, expectedDockerFileContents?: string): Promise<void> {
         let previousOs = ext.os;
         ext.os = {
             platform: hostOs === 'Windows' ? 'win32' : 'linux',